contendo_api/functions/get_top_lists/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
google-cloud-storage
google-cloud-error-reporting
pandas
contendo_utils>=0.2